+++ files/gitlab-cookbooks/gitlab-ee/recipes/default.rb
@@ -24,7 +24,8 @@ include_recipe 'gitlab::default'
   'sidekiq-cluster',
   'geo-postgresql',
   'geo-logcursor',
-  'pgbouncer'
+  'pgbouncer',
+  'pgbouncer-exporter'
 ].each do |service|
   if node['gitlab'][service]['enable']
     include_recipe "gitlab-ee::#{service}"