+++ CHANGELOG.md
@@ -13,6 +13,7 @@ omnibus-gitlab repository.
 - Upgrade Ruby version to 2.3.7
 - Update gitlab-monitor to 2.13.0
 - Add Prometheus Alertmanager
+- Add pgbouncer_exporter
 - Bump git to 2.16.3
 - Bump curl to 7.59.0 aab088263
 - Bump pcre to 8.42 07c83a5ab