+++ CHANGELOG.md
@@ -102,6 +102,7 @@ omnibus-gitlab repository.
 
 - Pages: if logformat set to json, do not append timestamps with svlogd.
 - Downgrade jemalloc to 4.2.1 to avoid segfaults in Ruby
+- Mattermost: Fix reconfiguration of GitLab OAuth configuration settings
 
 10.6.0