+++ CHANGELOG.md
@@ -8,6 +8,7 @@ omnibus-gitlab repository.
 - Disable PgBouncer PID file by default
 - Bump minimum version required for upgrade to 10.8 2df263267
 - Mattermost: Fix reconfiguration of GitLab OAuth configuration settings
+- Bump PgBouncer to 1.8.1
 
 10.8.0